refactor: 온보딩 레이아웃 버튼들 공용 버튼 컴포넌트로 수정
