+++ app/components/icon/index.ts
@@ -1,2 +1,3 @@
 export { default as IconEdit } from './icon-edit';
 export { default as IconPlus } from './icon-plus';
+export { default as IconNavigateLeft } from './icon-navigate-left';
