+++ app/components/index.ts
@@ -2,10 +2,12 @@ export { default as Checkbox } from './checkbox';
 export { default as Input } from './input';
 export { default as Button } from './ui/button';
 export { default as Dialog } from '../components/dialog';
+export { default as BottomSheet } from '../components/bottom-sheet';
 export { default as DialogLogout } from '../components/ui/dialog-logout';
 export { default as SigninBottom } from '../components/ui/signin-bottom';
 export { default as FloatingButton } from './ui/floating-button';
 export { default as SocialButton } from './ui/social-button';
+export { default as SocialBottomSheet } from './ui/social-bottom-sheet';
 export { default as ToastContainer } from './toast-container';
 
 export * from './icon';
